Started refactoring for GUI
