+++ userList.txt
@@ -2,3 +2,4 @@ Tomi - asd
 Balazs - asd
 Atyad - tomi
 Emese - 123
+balazs - szia
